% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_create_analyzer}
\alias{accessanalyzer_create_analyzer}
\title{Creates an analyzer for your account}
\usage{
accessanalyzer_create_analyzer(
  analyzerName,
  archiveRules = NULL,
  clientToken = NULL,
  tags = NULL,
  type
)
}
\arguments{
\item{analyzerName}{[required] The name of the analyzer to create.}

\item{archiveRules}{Specifies the archive rules to add for the analyzer. Archive rules
automatically archive findings that meet the criteria you define for the
rule.}

\item{clientToken}{A client token.}

\item{tags}{The tags to apply to the analyzer.}

\item{type}{[required] The type of analyzer to create. Only ACCOUNT and ORGANIZATION analyzers
are supported. You can create only one analyzer per account per Region.
You can create up to 5 analyzers per organization per Region.}
}
\description{
Creates an analyzer for your account.

See \url{https://paws-r.github.io/docs/accessanalyzer/create_analyzer.html} for full documentation.
}
\keyword{internal}
