% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessanalyzer_operations.R
\name{accessanalyzer_list_findings}
\alias{accessanalyzer_list_findings}
\title{Retrieves a list of findings generated by the specified analyzer}
\usage{
accessanalyzer_list_findings(
  analyzerArn,
  filter = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sort = NULL
)
}
\arguments{
\item{analyzerArn}{[required] The \href{https://docs.aws.amazon.com/IAM/latest/UserGuide/access-analyzer-getting-started.html#permission-resources}{ARN of the analyzer}
to retrieve findings from.}

\item{filter}{A filter to match for the findings to return.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token used for pagination of results returned.}

\item{sort}{The sort order for the findings returned.}
}
\description{
Retrieves a list of findings generated by the specified analyzer.

See \url{https://paws-r.github.io/docs/accessanalyzer/list_findings.html} for full documentation.
}
\keyword{internal}
