% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudhsm_operations.R
\name{cloudhsm_create_hapg}
\alias{cloudhsm_create_hapg}
\title{This is documentation for AWS CloudHSM Classic}
\usage{
cloudhsm_create_hapg(Label)
}
\arguments{
\item{Label}{[required] The label of the new high-availability partition group.}
}
\description{
This is documentation for \strong{AWS CloudHSM Classic}. For more information, see \href{https://aws.amazon.com/cloudhsm/faqs/}{AWS CloudHSM Classic FAQs}, the \href{https://docs.aws.amazon.com/cloudhsm/classic/userguide/}{AWS CloudHSM Classic User Guide}, and the \href{https://docs.aws.amazon.com/cloudhsm/classic/APIReference/}{AWS CloudHSM Classic API Reference}.

See \url{https://paws-r.github.io/docs/cloudhsm/create_hapg.html} for full documentation.
}
\keyword{internal}
