% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detective_operations.R
\name{detective_disable_organization_admin_account}
\alias{detective_disable_organization_admin_account}
\title{Removes the Detective administrator account in the current Region}
\usage{
detective_disable_organization_admin_account()
}
\description{
Removes the Detective administrator account in the current Region. Deletes the organization behavior graph.

See \url{https://paws-r.github.io/docs/detective/disable_organization_admin_account.html} for full documentation.
}
\keyword{internal}
