% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_enable_profile}
\alias{iamrolesanywhere_enable_profile}
\title{Enables the roles in a profile to receive session credentials in
CreateSession}
\usage{
iamrolesanywhere_enable_profile(profileId)
}
\arguments{
\item{profileId}{[required] The unique identifier of the profile.}
}
\description{
Enables the roles in a profile to receive session credentials in \href{https://docs.aws.amazon.com/rolesanywhere/latest/APIReference/}{CreateSession}.

See \url{https://paws-r.github.io/docs/iamrolesanywhere/enable_profile.html} for full documentation.
}
\keyword{internal}
