% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iamrolesanywhere_operations.R
\name{iamrolesanywhere_list_trust_anchors}
\alias{iamrolesanywhere_list_trust_anchors}
\title{Lists the trust anchors in the authenticated account and Amazon Web
Services Region}
\usage{
iamrolesanywhere_list_trust_anchors(nextToken = NULL, pageSize = NULL)
}
\arguments{
\item{nextToken}{A token that indicates where the output should continue from, if a
previous operation did not show all results. To get the next results,
call the operation again with this value.}

\item{pageSize}{The number of resources in the paginated list.}
}
\description{
Lists the trust anchors in the authenticated account and Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/iamrolesanywhere/list_trust_anchors.html} for full documentation.
}
\keyword{internal}
