% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_create_findings_report}
\alias{inspector2_create_findings_report}
\title{Creates a finding report}
\usage{
inspector2_create_findings_report(
  filterCriteria = NULL,
  reportFormat,
  s3Destination
)
}
\arguments{
\item{filterCriteria}{The filter criteria to apply to the results of the finding report.}

\item{reportFormat}{[required] The format to generate the report in.}

\item{s3Destination}{[required] The Amazon S3 export destination for the report.}
}
\description{
Creates a finding report.

See \url{https://paws-r.github.io/docs/inspector2/create_findings_report.html} for full documentation.
}
\keyword{internal}
