% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_disable_delegated_admin_account}
\alias{inspector2_disable_delegated_admin_account}
\title{Disables the Amazon Inspector delegated administrator for your
organization}
\usage{
inspector2_disable_delegated_admin_account(delegatedAdminAccountId)
}
\arguments{
\item{delegatedAdminAccountId}{[required] The Amazon Web Services account ID of the current Amazon Inspector
delegated administrator.}
}
\description{
Disables the Amazon Inspector delegated administrator for your organization.

See \url{https://paws-r.github.io/docs/inspector2/disable_delegated_admin_account.html} for full documentation.
}
\keyword{internal}
