% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_filters}
\alias{inspector2_list_filters}
\title{Lists the filters associated with your account}
\usage{
inspector2_list_filters(
  action = NULL,
  arns = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{action}{The action the filter applies to matched findings.}

\item{arns}{The Amazon resource number (ARN) of the filter.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}
}
\description{
Lists the filters associated with your account.

See \url{https://paws-r.github.io/docs/inspector2/list_filters.html} for full documentation.
}
\keyword{internal}
