% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_tags_for_resource}
\alias{inspector2_list_tags_for_resource}
\title{Lists all tags attached to a given resource}
\usage{
inspector2_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The Amazon resource number (ARN) of the resource to list tags of.}
}
\description{
Lists all tags attached to a given resource.

See \url{https://paws-r.github.io/docs/inspector2/list_tags_for_resource.html} for full documentation.
}
\keyword{internal}
