% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector2_operations.R
\name{inspector2_list_usage_totals}
\alias{inspector2_list_usage_totals}
\title{Lists the Amazon Inspector usage totals over the last 30 days}
\usage{
inspector2_list_usage_totals(
  accountIds = NULL,
  maxResults = NULL,
  nextToken = NULL
)
}
\arguments{
\item{accountIds}{The Amazon Web Services account IDs to retrieve usage totals for.}

\item{maxResults}{The maximum number of results to return in the response.}

\item{nextToken}{A token to use for paginating results that are returned in the response.
Set the value of this parameter to null for the first request to a list
action. For subsequent calls, use the \code{NextToken} value returned from
the previous request to continue listing results after the first page.}
}
\description{
Lists the Amazon Inspector usage totals over the last 30 days.

See \url{https://paws-r.github.io/docs/inspector2/list_usage_totals.html} for full documentation.
}
\keyword{internal}
