% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_register_cross_account_access_role}
\alias{inspector_register_cross_account_access_role}
\title{Registers the IAM role that grants Amazon Inspector access to AWS
Services needed to perform security assessments}
\usage{
inspector_register_cross_account_access_role(roleArn)
}
\arguments{
\item{roleArn}{[required] The ARN of the IAM role that grants Amazon Inspector access to AWS
Services needed to perform security assessments.}
}
\description{
Registers the IAM role that grants Amazon Inspector access to AWS Services needed to perform security assessments.

See \url{https://paws-r.github.io/docs/inspector/register_cross_account_access_role.html} for full documentation.
}
\keyword{internal}
