% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspector_operations.R
\name{inspector_update_assessment_target}
\alias{inspector_update_assessment_target}
\title{Updates the assessment target that is specified by the ARN of the
assessment target}
\usage{
inspector_update_assessment_target(
  assessmentTargetArn,
  assessmentTargetName,
  resourceGroupArn = NULL
)
}
\arguments{
\item{assessmentTargetArn}{[required] The ARN of the assessment target that you want to update.}

\item{assessmentTargetName}{[required] The name of the assessment target that you want to update.}

\item{resourceGroupArn}{The ARN of the resource group that is used to specify the new resource
group to associate with the assessment target.}
}
\description{
Updates the assessment target that is specified by the ARN of the assessment target.

See \url{https://paws-r.github.io/docs/inspector/update_assessment_target.html} for full documentation.
}
\keyword{internal}
