% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_describe_buckets}
\alias{macie2_describe_buckets}
\title{Retrieves (queries) statistical data and other information about one or
more S3 buckets that Amazon Macie monitors and analyzes}
\usage{
macie2_describe_buckets(
  criteria = NULL,
  maxResults = NULL,
  nextToken = NULL,
  sortCriteria = NULL
)
}
\arguments{
\item{criteria}{The criteria to use to filter the query results.}

\item{maxResults}{The maximum number of items to include in each page of the response. The
default value is 50.}

\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}

\item{sortCriteria}{The criteria to use to sort the query results.}
}
\description{
Retrieves (queries) statistical data and other information about one or more S3 buckets that Amazon Macie monitors and analyzes.

See \url{https://paws-r.github.io/docs/macie2/describe_buckets.html} for full documentation.
}
\keyword{internal}
