% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_findings_filter}
\alias{macie2_get_findings_filter}
\title{Retrieves the criteria and other settings for a findings filter}
\usage{
macie2_get_findings_filter(id)
}
\arguments{
\item{id}{[required] The unique identifier for the Amazon Macie resource or account that the
request applies to.}
}
\description{
Retrieves the criteria and other settings for a findings filter.

See \url{https://paws-r.github.io/docs/macie2/get_findings_filter.html} for full documentation.
}
\keyword{internal}
