% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_get_master_account}
\alias{macie2_get_master_account}
\title{(Deprecated) Retrieves information about the Amazon Macie administrator
account for an account}
\usage{
macie2_get_master_account()
}
\description{
(Deprecated) Retrieves information about the Amazon Macie administrator account for an account. This operation has been replaced by the GetAdministratorAccount operation.

See \url{https://paws-r.github.io/docs/macie2/get_master_account.html} for full documentation.
}
\keyword{internal}
