% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie2_operations.R
\name{macie2_list_managed_data_identifiers}
\alias{macie2_list_managed_data_identifiers}
\title{Retrieves information about all the managed data identifiers that Amazon
Macie currently provides}
\usage{
macie2_list_managed_data_identifiers(nextToken = NULL)
}
\arguments{
\item{nextToken}{The nextToken string that specifies which page of results to return in a
paginated response.}
}
\description{
Retrieves information about all the managed data identifiers that Amazon Macie currently provides.

See \url{https://paws-r.github.io/docs/macie2/list_managed_data_identifiers.html} for full documentation.
}
\keyword{internal}
