% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie_operations.R
\name{macie_associate_s3_resources}
\alias{macie_associate_s3_resources}
\title{(Discontinued) Associates specified S3 resources with Amazon Macie
Classic for monitoring and data classification}
\usage{
macie_associate_s3_resources(memberAccountId = NULL, s3Resources)
}
\arguments{
\item{memberAccountId}{(Discontinued) The ID of the Amazon Macie Classic member account whose
resources you want to associate with Macie Classic.}

\item{s3Resources}{[required] (Discontinued) The S3 resources that you want to associate with Amazon
Macie Classic for monitoring and data classification.}
}
\description{
(Discontinued) Associates specified S3 resources with Amazon Macie Classic for monitoring and data classification. If \code{memberAccountId} isn't specified, the action associates specified S3 resources with Macie Classic for the current Macie Classic administrator account. If \code{memberAccountId} is specified, the action associates specified S3 resources with Macie Classic for the specified member account.

See \url{https://paws-r.github.io/docs/macie/associate_s3_resources.html} for full documentation.
}
\keyword{internal}
