% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macie_operations.R
\name{macie_disassociate_s3_resources}
\alias{macie_disassociate_s3_resources}
\title{(Discontinued) Removes specified S3 resources from being monitored by
Amazon Macie Classic}
\usage{
macie_disassociate_s3_resources(memberAccountId = NULL, associatedS3Resources)
}
\arguments{
\item{memberAccountId}{(Discontinued) The ID of the Amazon Macie Classic member account whose
resources you want to remove from being monitored by Macie Classic.}

\item{associatedS3Resources}{[required] (Discontinued) The S3 resources (buckets or prefixes) that you want to
remove from being monitored and classified by Amazon Macie Classic.}
}
\description{
(Discontinued) Removes specified S3 resources from being monitored by Amazon Macie Classic. If \code{memberAccountId} isn't specified, the action removes specified S3 resources from Macie Classic for the current Macie Classic administrator account. If \code{memberAccountId} is specified, the action removes specified S3 resources from Macie Classic for the specified member account.

See \url{https://paws-r.github.io/docs/macie/disassociate_s3_resources.html} for full documentation.
}
\keyword{internal}
