% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ram_operations.R
\name{ram_untag_resource}
\alias{ram_untag_resource}
\title{Removes the specified tag key and value pairs from the specified
resource share}
\usage{
ram_untag_resource(resourceShareArn, tagKeys)
}
\arguments{
\item{resourceShareArn}{[required] Specifies the \href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{Amazon Resoure Name (ARN)}
of the resource share that you want to remove tags from. The tags are
removed from the resource share, not the resources in the resource
share.}

\item{tagKeys}{[required] Specifies a list of one or more tag keys that you want to remove.}
}
\description{
Removes the specified tag key and value pairs from the specified resource share.

See \url{https://paws-r.github.io/docs/ram/untag_resource.html} for full documentation.
}
\keyword{internal}
