% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_create_subscription}
\alias{shield_create_subscription}
\title{Activates Shield Advanced for an account}
\usage{
shield_create_subscription()
}
\description{
Activates Shield Advanced for an account.

See \url{https://paws-r.github.io/docs/shield/create_subscription.html} for full documentation.
}
\keyword{internal}
