% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield_operations.R
\name{shield_describe_emergency_contact_settings}
\alias{shield_describe_emergency_contact_settings}
\title{A list of email addresses and phone numbers that the Shield Response
Team (SRT) can use to contact you if you have proactive engagement
enabled, for escalations to the SRT and to initiate proactive customer
support}
\usage{
shield_describe_emergency_contact_settings()
}
\description{
A list of email addresses and phone numbers that the Shield Response Team (SRT) can use to contact you if you have proactive engagement enabled, for escalations to the SRT and to initiate proactive customer support.

See \url{https://paws-r.github.io/docs/shield/describe_emergency_contact_settings.html} for full documentation.
}
\keyword{internal}
