% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssoadmin_operations.R
\name{ssoadmin_describe_permission_set}
\alias{ssoadmin_describe_permission_set}
\title{Gets the details of the permission set}
\usage{
ssoadmin_describe_permission_set(InstanceArn, PermissionSetArn)
}
\arguments{
\item{InstanceArn}{[required] The ARN of the Amazon Web Services SSO instance under which the
operation will be executed. For more information about ARNs, see Amazon
Resource Names (ARNs) and Amazon Web Services Service Namespaces in the
\emph{Amazon Web Services General Reference}.}

\item{PermissionSetArn}{[required] The ARN of the permission set.}
}
\description{
Gets the details of the permission set.

See \url{https://paws-r.github.io/docs/ssoadmin/describe_permission_set.html} for full documentation.
}
\keyword{internal}
